- Improve `init` and `destroy`. - Reuse destroyed `ID`s. - Added support for `hash_first`, `hash_next` and `hash_last` functions. - Improved tests.

+++ wrapper_randomx_library/build-wrapper.sh
@@ -7,11 +7,24 @@ cmake .
 make
 
 if [ -f "libwrapper_randomx.1.0.0.dylib" ]; then
-	rm libwrapper_randomx.dylib
+
+	if [[ `uname -m` == 'arm64' ]]; then
+	  rm libwrapper_randomx.dylib
+    rm libwrapper_randomx-arm64.dylib
+
+    cp -f libwrapper_randomx.1.0.0.dylib libwrapper_randomx-arm64.dylib
+    cp -f libwrapper_randomx-x64.dylib libwrapper_randomx.dylib
+  else
+    rm libwrapper_randomx.dylib
+    rm libwrapper_randomx-x64.dylib
+
+    cp -f libwrapper_randomx.1.0.0.dylib libwrapper_randomx-x64.dylib
     cp -f libwrapper_randomx.1.0.0.dylib libwrapper_randomx.dylib
+  fi
+
 fi
 
 if [ -f "libwrapper_randomx.so.1.0.0" ]; then
 	rm libwrapper_randomx.so
-    cp -f libwrapper_randomx.so.1.0.0 libwrapper_randomx.so
+  cp -f libwrapper_randomx.so.1.0.0 libwrapper_randomx.so
 fi
